+++ src/main/java/io/github/thebusybiscuit/slimefun4/implementation/items/autocrafters/VanillaAutoCrafter.java
@@ -9,7 +9,6 @@ import javax.annotation.Nullable;
 import javax.annotation.ParametersAreNonnullByDefault;
 
 import org.apache.commons.lang.Validate;
-import org.bukkit.Bukkit;
 import org.bukkit.ChatColor;
 import org.bukkit.Material;
 import org.bukkit.NamespacedKey;
@@ -22,9 +21,11 @@ import org.bukkit.inventory.ItemStack;
 import org.bukkit.inventory.Recipe;
 import org.bukkit.inventory.ShapedRecipe;
 import org.bukkit.inventory.ShapelessRecipe;
+import org.bukkit.persistence.PersistentDataContainer;
+import org.bukkit.persistence.PersistentDataType;
 
-import io.github.thebusybiscuit.cscorelib2.data.PersistentDataAPI;
 import io.github.thebusybiscuit.cscorelib2.item.CustomItem;
+import io.github.thebusybiscuit.slimefun4.core.services.MinecraftRecipeService;
 import io.github.thebusybiscuit.slimefun4.implementation.SlimefunPlugin;
 import io.github.thebusybiscuit.slimefun4.implementation.tasks.AsyncRecipeChoiceTask;
 import io.github.thebusybiscuit.slimefun4.utils.ChestMenuUtils;
@@ -58,16 +59,12 @@ public class VanillaAutoCrafter extends AbstractAutoCrafter {
     @Override
     @Nullable
     public AbstractRecipe getSelectedRecipe(@Nonnull Block b) {
-        return AbstractRecipe.of(getRecipe(b));
-    }
-
-    @Nullable
-    private Recipe getRecipe(@Nonnull Block b) {
         BlockState state = PaperLib.getBlockState(b, false).getState();
 
         if (state instanceof Skull) {
             // Read the stored value from persistent data storage
-            String value = PersistentDataAPI.getString((Skull) state, recipeStorageKey);
+            PersistentDataContainer container = ((Skull) state).getPersistentDataContainer();
+            String value = container.get(recipeStorageKey, PersistentDataType.STRING);
 
             if (value != null) {
                 String[] values = PatternUtils.COLON.split(value);
@@ -79,8 +76,15 @@ public class VanillaAutoCrafter extends AbstractAutoCrafter {
                  */
                 @SuppressWarnings("deprecation")
                 NamespacedKey key = new NamespacedKey(values[0], values[1]);
+                Recipe keyedRecipe = SlimefunPlugin.getMinecraftRecipeService().getRecipe(key);
 
-                return SlimefunPlugin.getMinecraftRecipeService().getRecipe(key);
+                if (keyedRecipe != null) {
+                    boolean enabled = !container.has(recipeEnabledKey, PersistentDataType.BYTE);
+                    AbstractRecipe recipe = AbstractRecipe.of(keyedRecipe);
+                    recipe.setEnabled(enabled);
+
+                    return recipe;
+                }
             }
         }
 
@@ -169,7 +173,10 @@ public class VanillaAutoCrafter extends AbstractAutoCrafter {
     private List<Recipe> getRecipesFor(@Nonnull ItemStack item) {
         List<Recipe> recipes = new ArrayList<>();
 
-        for (Recipe recipe : Bukkit.getRecipesFor(item)) {
+        // Fixes #2913 - Bukkit.getRecipesFor() only checks for Materials
+        MinecraftRecipeService recipeService = SlimefunPlugin.getMinecraftRecipeService();
+
+        for (Recipe recipe : recipeService.getRecipesFor(item)) {
             if (recipe instanceof ShapedRecipe || recipe instanceof ShapelessRecipe) {
                 recipes.add(recipe);
             }